# api/api-docs.yaml
openapi: 3.0.0
info:
  title: GoMQTT API
  description: |
    REST API for managing GoMQTT, a modern MQTT broker.
    This API provides endpoints for managing clients, messages, users, permissions, and roles.
  version: 1.0.0
  contact:
    name: GoMQTT Team
    url: https://github.com/mstgnz/gomqtt
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/clients:
    get:
      summary: List clients
      description: Get a list of all connected MQTT clients
      tags:
        - Clients
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '401':
          description: Unauthorized
...
  /api/messages:
    get:
      summary: List messages
      description: Get a list of messages with optional filtering
      tags:
        - Messages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Client:
      type: object
      properties:
        id:
          type: string
          description: Client identifier
        username:
          type: string
          description: Client username
        connected_at:
          type: string
          format: date-time
          description: Connection timestamp
        ip_address:
          type: string
          description: Client IP address
        protocol:
          type: string
          description: MQTT protocol version
        keep_alive:
          type: integer
          description: Keep alive interval in seconds
        clean_session:
          type: boolean
          description: Whether the session is clean
        subscriptions:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              qos:
                type: integer
                enum: [0, 1, 2]
    Message:
      type: object
      properties:
        id:
          type: string
          description: Message identifier
        topic:
          type: string
          description: The topic
        payload:
          type: string
          description: Message payload
        qos:
          type: integer
          enum: [0, 1, 2]
          description: Quality of Service
        retained:
          type: boolean
          description: Whether the message is retained
        client_id:
          type: string
          description: ID of the client that published the message
        timestamp:
          type: string
          format: date-time
          description: When the message was published
...